set since tag for gpg stuff to 2.4.0, rename Signature.key to Signature.keyId and added DummyGPG for testing
